2nd chart

+++ backend/models/models.go
@@ -70,11 +70,11 @@ type GraphicsInfo struct {
 }
 
 type SalariesPie struct {
-	Sub5   float64 `json:"sub5"`
-	Sub10  float64 `json:"sub10"`
-	Sub15  float64 `json:"sub15"`
-	Sub20  float64 `json:"sub20"`
-	Over20 float64 `json:"over20"`
+	Sub5   int `json:"sub5"`
+	Sub10  int `json:"sub10"`
+	Sub15  int `json:"sub15"`
+	Sub20  int `json:"sub20"`
+	Over20 int `json:"over20"`
 }
 
 type ResponseDbInfo struct {
